Crate pallet_nodle_uniques

Handle the ability to notify other pallets that they should stop all

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_nodle_uniques

Type Aliases

BalanceOf