Function pallet_nodle_uniques::pallet::dispatchables::create

pub fn create<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    admin: <<T as Config>::Lookup as StaticLookup>::Source
)

Issue a new collection of non-fungible items from a public origin.

This new collection has no items initially and its owner is the origin.

The origin must conform to the configured CreateOrigin and have sufficient funds free.

ItemDeposit funds of sender are reserved.

Parameters:

collection: The identifier of the new collection. This must not be currently in use.

admin: The admin of this collection. The admin is the initial address of each member of the collection’s admin team.

Emits Created event when successful.

Weight: O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::create.