Function pallet_nodle_uniques::pallet::dispatchables::force_item_status

pub fn force_item_status<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    owner: <<T as Config>::Lookup as StaticLookup>::Source,
    issuer: <<T as Config>::Lookup as StaticLookup>::Source,
    admin: <<T as Config>::Lookup as StaticLookup>::Source,
    freezer: <<T as Config>::Lookup as StaticLookup>::Source,
    free_holding: bool,
    is_frozen: bool
)

Alter the attributes of a given item.

Origin must be ForceOrigin.

collection: The identifier of the item.

owner: The new Owner of this item.

issuer: The new Issuer of this item.

admin: The new Admin of this item.

freezer: The new Freezer of this item.

free_holding: Whether a deposit is taken for holding an item of this collection.

is_frozen: Whether this collection is frozen except for permissioned/admin instructions.

Emits ItemStatusChanged with the identity of the item.

Weight: O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::force_item_status.