Function pallet_nodle_uniques::pallet::dispatchables::set_collection_metadata

pub fn set_collection_metadata<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    data: BoundedVec<u8, T::StringLimit>,
    is_frozen: bool
)

Set the metadata for a collection.

Origin must be either ForceOrigin or Signed and the sender should be the Owner of the collection.

If the origin is Signed, then funds of signer are reserved according to the formula: MetadataDepositBase + DepositPerByte * data.len taking into account any already reserved funds.

collection: The identifier of the item whose metadata to update.

data: The general information of this item. Limited in length by StringLimit.

is_frozen: Whether the metadata should be frozen against further changes.

Emits CollectionMetadataSet.

Weight: O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_collection_metadata.