pallet_nodle_uniques::pallet::dispatchables
Function transfer_ownership

pub fn transfer_ownership<T: Config<I>, I: 'static>(
    collection: T::CollectionId,
    owner: <<T as Config>::Lookup as StaticLookup>::Source,
)

Change the Owner of a collection.

Origin must be Signed and the sender should be the Owner of the collection.

collection: The collection whose owner should be changed.

owner: The new Owner of this collection. They must have called set_accept_ownership with collection in order for this operation to succeed.

Emits OwnerChanged.

Weight: O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::transfer_ownership.