pallet_nodle_uniques::pallet
Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

approve_transfer

Approve an item to be transferred by a delegated third-party account.

burn

Destroy a single item.

buy_item

Allows to buy an item if it’s up for sale.

cancel_approval

Cancel the prior approval for the transfer of an item by a delegate.

clear_attribute

Clear an attribute for a collection or item.

clear_collection_metadata

Clear the metadata for a collection.

clear_metadata

Clear the metadata for an item.

create

Issue a new collection of non-fungible items from a public origin.

create_with_extra_deposit_limit

Issue a new collection of non-fungible items from a public origin.

destroy

Destroy a collection of fungible items.

force_create

Issue a new collection of non-fungible items from a privileged origin.

force_item_status

Alter the attributes of a given item.

freeze

Disallow further unprivileged transfer of an item.

freeze_collection

Disallow further unprivileged transfers for a whole collection.

mint

Mint an item of a particular collection.

mint_with_extra_deposit

Mint an item of a particular collection with extra deposit.

redeposit

Reevaluate the deposits on some items.

set_accept_ownership

Set (or reset) the acceptance of ownership for a particular account.

set_attribute

Set an attribute for a collection or item.

set_collection_max_supply

Set the maximum amount of items a collection could have.

set_collection_metadata

Set the metadata for a collection.

set_metadata

Set the metadata for an item.

set_price

Set (or reset) the price for an item.

set_team

Change the Issuer, Admin and Freezer of a collection.

thaw

Re-allow unprivileged transfer of an item.

thaw_collection

Re-allow unprivileged transfers for a whole collection.

transfer

Move an item from the sender account to another.

transfer_ownership

Change the Owner of a collection.

update_extra_deposit_limit

Update extra deposit limit for a collection when it’s not against what is already reserved.